Changed renderer to selection and make field required.
It was reported that users do not recognize all available templates in the
dropdown. So we changed this to a selection field with the length of available
templates.
As selections are not preselected on default (unlike to dropdowns) we need to
ensure that the user actually selects an entry by marking this field required.